www/Django/ScrumMaster/ScrumMaster/routing.py
from channels.auth import AuthMiddlewareStack
from channels.routing import ProtocolTypeRouter, URLRouter
import Scrum.routing

application = ProtocolTypeRouter({
    'websocket': AuthMiddlewareStack(
        URLRouter(
            Scrum.routing.websocket_urlpatterns
        )
    )
})