www/Django/ScrumMaster/requirements.txt
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
certifi==2018.4.16
chardet==3.0.4
constantly==15.1.0
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-cors-headers==2.2.0
django-iedge==0.1
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gitdb2==2.0.4
GitPython==2.1.11
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
linuxjobber==0.1
mysqlclient==1.3.12
oauthlib==2.1.0
packaging==17.1
pip-review==1.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
selenium==3.14.0
six==1.11.0
smmap2==2.0.4
style==1.1.0
txaio==18.8.1
update==0.0.1
urllib3==1.23
zope.interface==4.5.0
ARGS=""

ARGS="$ARGS --log-to-terminal"
ARGS="$ARGS --port 8080"

exec mod_wsgi-express start-server $ARGS wsgi/application
